requirements.txt
textstreamlit>=1.28.0
chempy>=0.8.1
pubchempy>=1.0.5